+++ src/model/AppDataManager/DataClasses/DisplaySettingsData.ts
@@ -1,16 +1,6 @@
 import { DataViewConfiguration, TerminalHeightMode } from 'src/model/Settings/DisplaySettings/DisplaySettings';
 
-export class DisplaySettingsDataV1 {
-  version = 1;
-  charSizePx = 14;
-  verticalRowPaddingPx = 5;
-  terminalWidthChars = 120;
-  scrollbackBufferSizeRows = 2000;
-  dataViewConfiguration = DataViewConfiguration.SINGLE_TERMINAL;
-}
-
-export class DisplaySettingsDataV2 {
-  version = 1;
+export class DisplaySettingsData {
   charSizePx = 14;
   verticalRowPaddingPx = 5;
   terminalWidthChars = 120;